Bridge: Move remaining JNI and Node store-based APIs over to bridge_fn
The exception is Java's SessionCipher_EncryptMessage, which
instantiates a different Java type based on the message type.
